# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATGT Lâm Đồng API",
    "version": "0.0.1",
    "description": "Executable API contracts. Runtime endpoints may be delivered by later tasks."
  },
  "servers": [{ "url": "http://localhost:3000" }],
  "components": {
    "schemas": {
      "CreateSos": {
        "type": "object",
        "properties": {
          "coordinateLongitude": {
            "type": "number",
            "minimum": -180,
            "maximum": 180
          },
          "coordinateLatitude": {
            "type": "number",
            "minimum": -90,
            "maximum": 90
          },
          "accuracyMeters": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          },
          "idempotencyKey": { "type": "string", "format": "uuid" },
          "incidentType": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{1,63}$"
          },
          "description": { "type": "string", "minLength": 1, "maxLength": 500 },
          "clientEventAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "coordinateLongitude",
          "coordinateLatitude",
          "accuracyMeters",
          "idempotencyKey",
          "incidentType",
          "clientEventAt"
        ],
        "additionalProperties": false
      },
      "SosAccepted": {
        "type": "object",
        "properties": {
          "publicCode": {
            "type": "string",
            "pattern": "^[0-9ABCDEFGHJKMNPQRSTVWXYZ]{12}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "RECEIVED",
              "IN_PROGRESS",
              "COMPLETED",
              "INSUFFICIENT_BASIS"
            ]
          },
          "receivedAt": { "type": "string", "format": "date-time" },
          "emergencyContacts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "minLength": 1 },
                "number": {
                  "type": "string",
                  "enum": ["112", "113", "114", "115"]
                },
                "type": {
                  "type": "string",
                  "enum": ["112", "113", "114", "115"]
                }
              },
              "required": ["name", "number", "type"],
              "additionalProperties": false
            },
            "minItems": 1
          }
        },
        "required": ["publicCode", "status", "receivedAt", "emergencyContacts"],
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "minLength": 1 },
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string", "minLength": 1, "maxLength": 1000 },
          "instance": { "type": "string", "minLength": 1 },
          "error_code": {
            "type": "string",
            "enum": [
              "ATGT_VALIDATION_FAILED",
              "ATGT_COORDINATE_INVALID",
              "ATGT_UNAUTHORIZED",
              "ATGT_FORBIDDEN",
              "ATGT_AUTH_TOKEN_INVALID",
              "ATGT_AUTH_SESSION_REVOKED",
              "ATGT_AUTH_POLICY_MISSING",
              "ATGT_CITIZEN_SESSION_INVALID",
              "ATGT_CITIZEN_SESSION_EXPIRED",
              "ATGT_CITIZEN_SESSION_REVOKED",
              "ATGT_NOT_FOUND",
              "ATGT_CONFLICT",
              "ATGT_IDEMPOTENCY_CONFLICT",
              "ATGT_IDEMPOTENCY_IN_PROGRESS",
              "ATGT_INVALID_STATE_TRANSITION",
              "ATGT_INSUFFICIENT_PERMISSION_FOR_TRANSITION",
              "ATGT_EVIDENCE_MIME_NOT_ALLOWED",
              "ATGT_EVIDENCE_SIZE_EXCEEDED",
              "ATGT_EVIDENCE_HASH_MISMATCH",
              "ATGT_RATE_LIMIT_EXCEEDED",
              "ATGT_MAP_PROVIDER_UNAVAILABLE",
              "ATGT_INTERNAL_ERROR"
            ]
          },
          "trace_id": { "type": "string", "pattern": "^[0-9a-f]{32}$" }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "error_code",
          "trace_id"
        ],
        "additionalProperties": false
      },
      "CreateCitizenSession": {
        "type": "object",
        "properties": {
          "device_class": { "type": "string", "enum": ["mobile", "web"] }
        },
        "required": ["device_class"],
        "additionalProperties": false
      },
      "CitizenSession": {
        "type": "object",
        "properties": {
          "session_id": { "type": "string", "format": "uuid" },
          "session_token": {
            "type": "string",
            "minLength": 43,
            "maxLength": 128
          },
          "device_class": { "type": "string", "enum": ["mobile", "web"] },
          "created_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "rotate_after": { "type": "string", "format": "date-time" }
        },
        "required": [
          "session_id",
          "session_token",
          "device_class",
          "created_at",
          "expires_at",
          "rotate_after"
        ],
        "additionalProperties": false
      },
      "CitizenSessionHeader": {
        "type": "object",
        "properties": {
          "x-citizen-session": {
            "type": "string",
            "minLength": 43,
            "maxLength": 128
          }
        },
        "required": ["x-citizen-session"],
        "additionalProperties": false
      },
      "EventEnvelope": {
        "type": "object",
        "properties": {
          "event_id": { "type": "string", "format": "uuid" },
          "type": {
            "type": "string",
            "enum": [
              "incident.received.v1",
              "incident.verified.v1",
              "assignment.created.v1",
              "assignment.ack_timed_out.v1",
              "evidence.scan_requested.v1",
              "evidence.ready.v1",
              "map.version.published.v1",
              "notification.requested.v1",
              "provider.quota_threshold.v1"
            ]
          },
          "version": { "type": "integer", "exclusiveMinimum": 0 },
          "occurred_at": { "type": "string", "format": "date-time" },
          "trace_id": { "type": "string", "pattern": "^[0-9a-f]{32}$" },
          "aggregate_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "aggregate_type": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "data": { "type": "object", "additionalProperties": {} }
        },
        "required": [
          "event_id",
          "type",
          "version",
          "occurred_at",
          "trace_id",
          "aggregate_id",
          "aggregate_type",
          "data"
        ],
        "additionalProperties": false
      },
      "ProviderQuality": {
        "type": "string",
        "enum": ["LIVE", "CACHED", "DEGRADED", "UNAVAILABLE"]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/api/v1/incidents/sos": {
      "post": {
        "summary": "Submit an SOS incident",
        "description": "Contract-first endpoint scheduled for implementation in T07.",
        "tags": ["Incidents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSos" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "SOS accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SosAccepted" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/sessions": {
      "post": {
        "summary": "Create an anonymous citizen session",
        "tags": ["Identity"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCitizenSession" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Anonymous session created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CitizenSession" }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/sessions/rotate": {
      "post": {
        "summary": "Rotate an anonymous citizen session",
        "tags": ["Identity"],
        "parameters": [
          {
            "schema": { "type": "string", "minLength": 43, "maxLength": 128 },
            "required": true,
            "name": "x-citizen-session",
            "in": "header"
          }
        ],
        "responses": {
          "201": {
            "description": "Replacement session created and previous session revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CitizenSession" }
              }
            }
          },
          "401": {
            "description": "Session expired, revoked, or invalid",
            "content": {
              "application/problem+json": {
                "schema": { "$ref": "#/components/schemas/ProblemDetails" }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {}
}